Various small fixes: add ability to scroll back in time on chart; better coloring for heatmap

+++ include/2fish/models/orderbook_snapshot.h
@@ -7,7 +7,7 @@ struct OrderbookSnapshot {
 	std::array<double, 101> bids_{};
 	std::array<double, 101> asks_{};
 
-	int64_t timestamp_{}; // TODO: think about how to populate you
+	int64_t timestamp_{};
 
-	float getLiquidity(std::size_t level) const { return static_cast<float>(bids_[level] + asks_[level]); }
+	double getLiquidity(std::size_t level) const { return bids_[level] + asks_[level]; }
 };
